# By Igor Solovey (236) and others
# Via Igor Solovey (185) and others
* commit '76060fd5a6985e4a831d4802b5d411708239a8d4 [formerly 3a4725e60c315b1581c3f4fe7792290d9e71516c]': (1370 commits)
  no message
  [DF-8411] fixed crash in shadow
  [DF-8362] extra include removed
  [DF-8362] build for ios fixed
  [DF-8362] screen blink fixed
  build for mac fixed
  fix assert and render loop
  DF-7517 fix after merge with development
  DF-185 mem leak fixed
  DF-185 fixed texture quality
  [DF-8363] Merged code into re-stable
  [Df-8409] Fixed save entity as
  [DF-8233] Disable resource sharing
  [DF-8286] detail from review fixed
  DF-8397 fixed
  [DF-185] "Newer version of Win10 SDK detected" warning fix
  [DF-8110] styles: added hDynamicInLinePadding property
  [DF-8110] styles: fixed size policy default value
  [DF-8108] IsHIDConnected changed for Android, Mac, Ios
  [DF-8286] details from review fixed
  ...


Former-commit-id: 0252df095bf367574db55802647dbf4bb652d9c3

+++ Tools/ResourceEditor/Classes/Qt/Scene/System/OwnersSignatureSystem.cpp
@@ -59,9 +59,9 @@ void OwnersSignatureSystem::ProcessCommand(const Command2 *command, bool redo)
 
 bool OwnersSignatureSystem::IsCommandIdValid(int _id)
 {
-	for (int i = 0; i < COUNT_OF(validIDs); ++i)
-	{
-		if(validIDs[i] == _id)
+    for (size_t i = 0; i < COUNT_OF(validIDs); ++i)
+    {
+        if(validIDs[i] == _id)
 		{
 			return true;
 		}
